Split hooks/use-bounty-application.ts into domain files. Closes #276
