+++ hooks/use-application-contracts.ts
@@ -1,0 +1,61 @@
+"use client";
+
+export type ApplicationContractClient = {
+  apply: (params: {
+    applicant: string;
+    bountyId: bigint;
+    proposal: string;
+  }) => Promise<{ txHash: string }>;
+  selectApplicant: (params: {
+    creator: string;
+    bountyId: bigint;
+    applicant: string;
+  }) => Promise<{ txHash: string }>;
+  submitWork: (params: {
+    contributor: string;
+    bountyId: bigint;
+    workCid: string;
+  }) => Promise<{ txHash: string }>;
+  approveSubmission: (params: {
+    creator: string;
+    bountyId: bigint;
+    points: number;
+  }) => Promise<{ txHash: string }>;
+  applyForSlot: (params: {
+    applicant: string;
+    bountyId: bigint;
+  }) => Promise<{ txHash: string }>;
+};
+
+export type ApplicationErrorCode =
+  | "missing_contract_bindings"
+  | "already_applied"
+  | "tx_failed";
+
+export class ApplicationError extends Error {
+  code: ApplicationErrorCode;
+  constructor(code: ApplicationErrorCode, message: string) {
+    super(message);
+    this.code = code;
+  }
+}
+
+export function toBountyIdBigInt(id: string): bigint {
+  if (/^\d+$/.test(id)) return BigInt(id);
+  const hex = id.replace(/-/g, "");
+  if (/^[0-9a-f]+$/i.test(hex)) return BigInt(`0x${hex}`);
+  throw new ApplicationError("tx_failed", `Invalid bounty ID: "${id}"`);
+}
+
+export function resolveApplicationClient(): ApplicationContractClient {
+  const client = (
+    globalThis as { __applicationContracts?: ApplicationContractClient }
+  ).__applicationContracts;
+  if (!client) {
+    throw new ApplicationError(
+      "missing_contract_bindings",
+      "Application contract bindings unavailable. Ensure bindings are loaded.",
+    );
+  }
+  return client;
+}
